val_P
Added val_P addon to the build list
It is a TIOW valkyrie tweak addon
Modified it to actually apply on top of the base TIOW valkyrie addon as
a separate entity instead of a duplicate.
This seems to fix the issues encountered with placing some of the
valkyries in zeus.